backend/requirements.txt
aiofiles==23.2.1
alembic==1.12.1
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.10.4
asyncpg==0.31.0
bcrypt==4.0.1
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.104.1
greenlet==3.3.1
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.2
jiter==0.13.0
kombu==5.6.2
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.4.2
openai==2.20.0
packaging==26.0
passlib==1.7.4
pdfminer.six==20251230
pdfplumber==0.11.9
pgvector==0.4.2
Pillow==10.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pyasn1==0.6.2
pycparser==3.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pypdfium2==5.4.0
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.2.1
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
reportlab==4.0.7
rsa==4.9.1
sendgrid==6.11.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starkbank-ecdsa==2.2.0
starlette==0.27.0
slowapi==0.1.9
tenacity==8.2.3
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uvicorn==0.24.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
